requirements.txt
pandas>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
tqdm>=4.65.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6 